+++ src/BillingServiceProvider.php
@@ -14,7 +14,7 @@ class BillingServiceProvider extends ServiceProvider
     public function boot(): void
     {
         $this->publishes([
-            __DIR__ . '/../config/payfast.php' => config_path('payfast.php'),
+            __DIR__ . '/../config/billing.php' => config_path('billing.php'),
         ], 'config');
 
         $this->publishes([
@@ -39,25 +39,25 @@ class BillingServiceProvider extends ServiceProvider
     public function register()
     {
         $this->mergeConfigFrom(
-            __DIR__ . '/../config/payfast.php',
-            'payfast'
+            __DIR__ . '/../config/billing.php',
+            'billing'
         );
 
         $this->app->bind('payfast', function () {
             return new Payfast([
-                'merchant_id' => config('payfast.merchant_id'),
-                'merchant_key' => config('payfast.merchant_key'),
-                'passphrase' => config('payfast.passphrase'),
+                'merchant_id' => config('billing.payfast.merchant_id'),
+                'merchant_key' => config('billing.payfast.merchant_key'),
+                'passphrase' => config('billing.payfast.passphrase'),
 
-                'test_mode' => config('payfast.test_mode'),
+                'test_mode' => config('billing.payfast.test_mode'),
 
-                'merchant_id_test' => config('payfast.merchant_id_test'),
-                'merchant_key_test' => config('payfast.merchant_key_test'),
-                'passphrase_test' => config('payfast.passphrase_test'),
+                'merchant_id_test' => config('billing.payfast.merchant_id_test'),
+                'merchant_key_test' => config('billing.payfast.merchant_key_test'),
+                'passphrase_test' => config('billing.payfast.passphrase_test'),
 
-                'return_url' => config('payfast.return_url'),
-                'cancel_url' => config('payfast.cancel_url'),
-                'notify_url' => config('payfast.notify_url'),
+                'return_url' => config('billing.payfast.return_url'),
+                'cancel_url' => config('billing.payfast.cancel_url'),
+                'notify_url' => config('billing.payfast.notify_url'),
             ]);
         });
     }